Fix dependency vulnerability, may break reqs
This vulnerability in PyYaml was automatically caught by GitHub. I got really tired of the automated security alert emails, so I just upgraded to the lowest known secure version. Hopefully this does not break the package functionality. If it does, we'll just have to be sure that we upgrade all API calls for PyYAML to respect the new version.